clean slate: restart main.js from AS example code
Rather than starting at the top of NEP-4, Peter suggested starting with
the AssemblyScript example code and implementing those functions first.

This copies them into main.ts mostly as-is (with some corrections to
syntax and modifications so that it compiles), and copies the same info
into the test file with requirements given as individual test cases.
These tests are all stubbed out for now.

Note that it is impossible to implement only these functions, as they
assume that tokens have already been minted and are already owned by
various parties.